requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
requests>=2.28.0
beautifulsoup4==4.10.0
APScheduler==3.8.1
pytz==2022.1
python-dotenv==0.19.1 